phase-04-data-state/requirements.txt
flask==3.0.3
psycopg2-binary==2.9.9



